Update 1.2.5 - 7 Bag rule
Added the modern Tetris rule called the 7 bag rule. The 7 bag rule means that pieces are randomly generated into a bag of 7 pieces where all of the pieces are unique.
There are 2 bags, the current bag, and the next bag. Once the current bag is empty, it gets filled with the pieces from the next bag and the next bag's pieces are re-generated.